# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/anime/{slug}/download-jobs": {
      "post": {
        "operationId": "DownloadJobsController_create",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDownloadJobDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadJobReceiptResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "La solicitud no supera la validación.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "El recurso solicitado no existe.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Se superó temporalmente el límite de solicitudes.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "El servidor no pudo completar la operación.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "503": {
            "description": "La dependencia necesaria no está disponible.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Crea un trabajo durable para una serie o rango",
        "tags": ["download jobs"]
      }
    },
    "/api/v1/download-jobs/{id}": {
      "get": {
        "operationId": "DownloadJobsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadJobResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "La capacidad falta, no coincide o ha expirado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Se superó temporalmente el límite de solicitudes.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "El servidor no pudo completar la operación.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jobCapability": []
          }
        ],
        "summary": "Consulta progreso y resultados mediante capacidad",
        "tags": ["download jobs"]
      }
    },
    "/api/v1/download-jobs/{id}/retry": {
      "post": {
        "operationId": "DownloadJobsController_retry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadJobResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "La solicitud no supera la validación.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "La capacidad falta, no coincide o ha expirado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Se superó temporalmente el límite de solicitudes.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "El servidor no pudo completar la operación.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jobCapability": []
          }
        ],
        "summary": "Reintenta únicamente episodios fallidos",
        "tags": ["download jobs"]
      }
    },
    "/api/v1/download-jobs/{id}/cancel": {
      "post": {
        "operationId": "DownloadJobsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadJobResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "La capacidad falta, no coincide o ha expirado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Se superó temporalmente el límite de solicitudes.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "El servidor no pudo completar la operación.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jobCapability": []
          }
        ],
        "summary": "Cancela los ítems pendientes del trabajo",
        "tags": ["download jobs"]
      }
    }
  },
  "info": {
    "title": "AnimeHub API",
    "description": "Contrato público REST compartido por AnimeHub Web y Desktop.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://animehub-api.duardo.dev",
      "description": "Producción"
    },
    {
      "url": "http://localhost:8000",
      "description": "Desarrollo local"
    }
  ],
  "components": {
    "securitySchemes": {
      "jobCapability": {
        "scheme": "bearer",
        "bearerFormat": "opaque capability",
        "type": "http",
        "description": "Capacidad opaca y temporal devuelta al crear un trabajo de descarga; no es un JWT."
      }
    },
    "schemas": {
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "number"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": ["type", "title", "status", "detail", "instance"]
      },
...
      "ResolvedLinkDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": ["MEGA", "PIXELDRAIN", "MP4UPLOAD", "ONE_FICHIER"]
          },
          "url": {
            "type": "string"
          }
        },
        "required": ["provider", "url"]
      },
      "ResolvedEpisodeDto": {
        "type": "object",
        "properties": {
          "episodeNumber": {
            "type": "number"
          },
          "audio": {
            "type": "string",
            "enum": ["SUB", "DUB"]
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResolvedLinkDto"
            }
          },
          "errorCode": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["episodeNumber", "audio", "links"]
      },
...
      "CreateDownloadJobDto": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string",
            "enum": ["ALL", "RANGE"]
          },
          "audio": {
            "type": "string",
            "enum": ["SUB", "DUB"]
          },
          "providers": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["MEGA", "PIXELDRAIN", "MP4UPLOAD", "ONE_FICHIER"]
            }
          },
          "from": {
            "type": "number"
          },
          "to": {
            "type": "number"
          }
        },
        "required": ["scope", "audio", "providers"]
      },
      "DownloadJobReceiptDto": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": ["jobId", "accessToken", "expiresAt"]
      },
      "DownloadJobReceiptResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DownloadJobReceiptDto"
          }
        },
        "required": ["data"]
      },
      "DownloadJobDataDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "COMPLETED",
              "PARTIAL",
              "FAILED",
              "CANCELLED"
            ]
          },
          "packageName": {
            "type": "string"
          },
          "totalItems": {
            "type": "number"
          },
          "completedItems": {
            "type": "number"
          },
          "failedItems": {
            "type": "number"
          },
          "expiresAt": {
            "type": "string"
          },
          "episodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResolvedEpisodeDto"
            }
          }
        },
        "required": [
          "id",
          "status",
          "packageName",
          "totalItems",
          "completedItems",
          "failedItems",
          "expiresAt",
          "episodes"
        ]
      },
      "DownloadJobResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DownloadJobDataDto"
          }
        },
        "required": ["data"]
      }
    }
  }
}